[SP-2457][PDI-14882] Hang in OSGIPluginTracker caused by concurrent hash map access
Resolves two issues:
Main issue is the concurrent WeakHashMap access which causes a hang.
Second issue is that the objects in the WeakHashMap will not be released because the value of the WeakHashMap is not a weak reference. We are storing each key as a value in a corresponding WeakHashMap. For example the key object in instanceToReferenceMap is then stored as a value in referenceToInstanceMap (and vice versa) which would prevent the WeakHashMap from automatically removing keys from either.

To fix both problems I have switched the WeakHashMap to Guava ConcurrentMap with weak values and weak keys.